requirements.txt
streamlit>=1.20
Pillow>=9.0
